// app/core/interface/types.d.ts
export interface ActiveTypeProps  {
    id: number; 
    type_name: string;
    type_value?: string;
}

export interface ActiveTypeBooleanProps  {
    id: number; 
    type_name: string;
    type_value?: boolean;
}

export interface ProductTypeProps  {
    id: number; 
    name: string;
    value?: string;
}